app/requirements.txt
backports.tarfile==1.2.0
cupy-cuda13x==13.6.0
cython==3.1.3
dill==0.4.0
importlib-metadata==8.0.0
jaraco.collections==5.1.0
matplotlib==3.10.6
pip-chill==1.0.3
platformdirs==4.2.2
py-cpuinfo==9.0.0
pyfftw==0.15.0
scipy==1.16.1
sympy==1.14.0
tomli==2.0.1
tqdm==4.67.1
